made the basic test more robust, prepare for 0.04 release
there wree a bunch of fails from cpantesters where the first buffer read
was empty, then the test of the second buffer read contained the first
expected result. Therefore it seems that the promise was resolving
sooner than we could test the server's result.

Instead make the test server emit an even once it has read, then we use
that to stop the ioloop. We now then must independently test the promise
from the writer.